requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
checksumdir==1.2.0
click==8.1.6
command-not-found==0.3
contourpy==1.1.0
cryptography==3.4.8
cycler==0.11.0
dbus-python==1.2.18
distro==1.7.0
distro-info===1.1build1
dnspython==2.4.1
email-validator==2.0.0.post2
Flask==2.3.2
Flask-Login==0.6.2
flask-mailman==0.3.0
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.3.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
fonttools==4.41.0
greenlet==2.0.2
gyp==0.1
httplib2==0.20.2
idna==3.4
importlib-metadata==4.6.4
importlib-resources==6.0.0
itsdangerous==2.1.2
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
kiwisolver==1.4.4
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==2.1.3
matplotlib==3.7.2
mkdocs-material-extensions==1.1.1
more-itertools==8.10.0
mpi4py==3.1.3
netifaces==0.11.0
numpy==1.25.1
oauthlib==3.2.0
packaging==23.1
pandas==2.0.3
passlib==1.7.4
Pillow==10.0.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
python-apt==2.4.0+ubuntu1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==5.4.1
SecretStorage==3.3.1
six==1.16.0
SQLAlchemy==2.0.19
systemd-python==234
tk==0.1.0
typing_extensions==4.7.1
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
wadllib==1.3.6
Werkzeug==2.3.6
WTForms==3.0.1
zipp==1.0.0
